+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Emerald Heart chat function only worked if Emerald Heart was DISABLED, instead of ENABLED. Fixed that.
 - Improved command response messages, for example "Added Claws to Dev" instead of "Added Claws to 1 entities".
 - Efficiency Attribute now uses modifiers instead of modifying the attribute base value. No idea why I was doing that originally, that's awful.
+- Support Slime now only checks if it should despawn once every 40 ticks rather than every tick, which should improve performance.